backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.1
email-validator==2.1.1
typing-extensions==4.9.0 
